refactor(connect): move tezos method definitions to corresponding folder
