+++ packages/connect-common/src/types/api/tezos/tezosGetPublicKey.ts
@@ -1,4 +1,4 @@
-import type { BundledParams, GetPublicKey, Params, PublicKey, Response } from '../params';
+import type { BundledParams, GetPublicKey, Params, PublicKey, Response } from '../../params';
 
 export declare function tezosGetPublicKey(params: Params<GetPublicKey>): Response<PublicKey>;
 export declare function tezosGetPublicKey(
