+++ packages/connect-common/src/types/api/tezos/tezosSignTransaction.ts
@@ -1,7 +1,7 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
-import type { TezosSignTransaction } from './tezos';
+import type { TezosSignTransaction } from './common';
+import type { Params, Response } from '../../params';
 
 export declare function tezosSignTransaction(
     params: Params<TezosSignTransaction>,
